Change var declaration to let
when you declare a variable with let, it's also hoisted to the top of its block scope (block-scoped variables), but unlike var,
it's not initialized until the actual declaration statement is encountered in the code. This is known as the "temporal dead zone", during which accessing the variable will result in a ReferenceError.

By declarig copy with let and initializing it at the same time we fix the issue