Check base case
We check if the value is of type of object or is null, if the condition is true, it means that the current level is not an object but a primitive value, so we return
true if it is the searched value ( we found it! ) and false it it is not.